requirements.txt
alembic==1.16.5
bcrypt==4.3.0
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.2
greenlet==3.2.4
iniconfig==2.1.0
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.3
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.3
shiboken6==6.9.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
